+++ src/main/java/com/owczarczak/footballers/match/MatchRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface MatchRepository extends JpaRepository<Match, Integer> {
+public interface MatchRepository extends JpaRepository<Match, Long> {
     @Query("SELECT m FROM Match m WHERE m.nameOfReferee =:name")
     List<Match> findByRefereeName(@Param("name") String name);
 }
